requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
PyPDF2==3.0.1
numpy==1.21.2
nltk==3.6.3
scikit-learn==0.24.2
google-generativeai>=0.3.0
gunicorn==20.1.0
Werkzeug==2.0.1 